Scenes: rename files to ScenesManagementCluster to prepare for code driven conversion.
